requirements.txt
speechrecognition
pyttsx3
pyaudio
pillow 